backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
sqlalchemy[asyncio]==2.0.34
asyncpg==0.29.0
pydantic-settings==2.5.2
python-multipart==0.0.9
fpdf2==2.7.9
celery==5.4.0
redis==5.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
